[DT-1063] Create payments arch module to hold domain classes
