+++ payments/base/arch/build.gradle.kts
@@ -1,0 +1,11 @@
+plugins {
+  id(GradlePluginId.ANDROID_LIBRARY)
+  id(GradlePluginId.ANDROID_MODULE)
+}
+
+android {
+  namespace = "com.appcoins.payments.arch"
+}
+
+dependencies {
+}
