+++ payments/base/arch/src/main/java/com/appcoins/payments/arch/PaymentMethodFactory.kt
@@ -1,0 +1,11 @@
+package com.appcoins.payments.arch
+
+interface PaymentMethodFactory<T> {
+  fun create(
+    wallet: WalletData,
+    developerWallet: String,
+    productInfo: ProductInfoData,
+    paymentMethodData: PaymentMethodData,
+    purchaseRequest: PurchaseRequest,
+  ): PaymentMethod<T>?
+}
