+++ payments/base/arch/src/main/java/com/appcoins/payments/arch/WalletProvider.kt
@@ -1,4 +1,4 @@
-package com.appcoins.payment_manager.wallet
+package com.appcoins.payments.arch
 
 interface WalletProvider {
