+++ payments/guest-wallet/src/main/java/com/appcoins/guest_wallet/RealWalletProvider.kt
@@ -2,8 +2,8 @@ package com.appcoins.guest_wallet
 
 import com.appcoins.guest_wallet.repository.WalletRepository
 import com.appcoins.guest_wallet.unique_id.UniqueIDProvider
-import com.appcoins.payment_manager.wallet.WalletData
-import com.appcoins.payment_manager.wallet.WalletProvider
+import com.appcoins.payments.arch.WalletData
+import com.appcoins.payments.arch.WalletProvider
 import javax.inject.Inject
 import javax.inject.Singleton
