+++ payments/osp-handler/src/main/java/com/appcoins/osp_handler/PaymentScreenContentProvider.kt
@@ -3,7 +3,7 @@ package com.appcoins.osp_handler
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import androidx.compose.runtime.Composable
-import com.appcoins.payment_manager.manager.PurchaseRequest
+import com.appcoins.payments.arch.PurchaseRequest
 
 interface PaymentScreenContentProvider {
   fun handleIntent(context: AppCompatActivity, intent: Intent?)
