+++ payments/payment-methods/adyen/src/main/java/com/appcoins/payment_method/adyen/CreditCardPaymentMethodFactory.kt
@@ -1,13 +1,13 @@
 package com.appcoins.payment_method.adyen
 
 import com.adyen.checkout.components.model.payments.request.PaymentMethodDetails
-import com.appcoins.payment_manager.manager.PurchaseRequest
-import com.appcoins.payment_manager.payment.PaymentMethod
-import com.appcoins.payment_manager.payment.PaymentMethodFactory
-import com.appcoins.payment_manager.repository.broker.domain.PaymentMethodData
-import com.appcoins.payment_manager.wallet.WalletData
 import com.appcoins.payment_method.adyen.repository.AdyenV2Repository
-import com.appcoins.product_inventory.domain.ProductInfoData
+import com.appcoins.payments.arch.PaymentMethod
+import com.appcoins.payments.arch.PaymentMethodData
+import com.appcoins.payments.arch.PaymentMethodFactory
+import com.appcoins.payments.arch.ProductInfoData
+import com.appcoins.payments.arch.PurchaseRequest
+import com.appcoins.payments.arch.WalletData
 import javax.inject.Inject
 import javax.inject.Singleton
