+++ payments/payment-methods/adyen/src/main/java/com/appcoins/payment_method/adyen/presentation/AdyenCreditCardScreenUiState.kt
@@ -7,8 +7,8 @@ import com.adyen.checkout.components.ActionComponentData
 import com.adyen.checkout.components.model.payments.response.Action
 import com.adyen.checkout.components.model.payments.response.RedirectAction
 import com.adyen.checkout.redirect.RedirectConfiguration
-import com.appcoins.payment_manager.manager.PurchaseRequest
-import com.appcoins.product_inventory.domain.ProductInfoData
+import com.appcoins.payments.arch.ProductInfoData
+import com.appcoins.payments.arch.PurchaseRequest
 
 sealed class AdyenCreditCardScreenUiState {
   object MakingPurchase : AdyenCreditCardScreenUiState()
