+++ payments/payment-methods/adyen/src/main/java/com/appcoins/payment_method/adyen/presentation/AdyenCreditCardViewModel.kt
@@ -20,14 +20,14 @@ import com.adyen.checkout.components.model.payments.response.Threeds2ChallengeAc
 import com.adyen.checkout.components.model.payments.response.Threeds2FingerprintAction
 import com.adyen.checkout.redirect.RedirectConfiguration
 import com.appcoins.payment_manager.manager.PaymentManager
-import com.appcoins.payment_manager.transaction.TransactionStatus.COMPLETED
-import com.appcoins.payment_manager.transaction.TransactionStatus.PENDING_SERVICE_AUTHORIZATION
-import com.appcoins.payment_manager.transaction.TransactionStatus.PENDING_USER_PAYMENT
-import com.appcoins.payment_manager.transaction.TransactionStatus.PROCESSING
-import com.appcoins.payment_manager.transaction.TransactionStatus.SETTLED
 import com.appcoins.payment_method.adyen.CreditCardPaymentMethod
 import com.appcoins.payment_method.adyen.CreditCardTransaction
 import com.appcoins.payment_prefs.domain.PreSelectedPaymentUseCase
+import com.appcoins.payments.arch.TransactionStatus.COMPLETED
+import com.appcoins.payments.arch.TransactionStatus.PENDING_SERVICE_AUTHORIZATION
+import com.appcoins.payments.arch.TransactionStatus.PENDING_USER_PAYMENT
+import com.appcoins.payments.arch.TransactionStatus.PROCESSING
+import com.appcoins.payments.arch.TransactionStatus.SETTLED
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.SharingStarted
