+++ payments/payment-methods/paypal/src/main/java/com/appcoins/payment_method/paypal/PaypalPaymentMethod.kt
@@ -1,12 +1,12 @@
 package com.appcoins.payment_method.paypal
 
-import com.appcoins.payment_manager.manager.PurchaseRequest
-import com.appcoins.payment_manager.payment.PaymentMethod
-import com.appcoins.payment_manager.wallet.WalletData
 import com.appcoins.payment_method.paypal.model.PaypalTransaction
 import com.appcoins.payment_method.paypal.repository.PaypalRepository
 import com.appcoins.payment_method.paypal.repository.model.PaymentDetailsRequest
-import com.appcoins.product_inventory.domain.ProductInfoData
+import com.appcoins.payments.arch.PaymentMethod
+import com.appcoins.payments.arch.ProductInfoData
+import com.appcoins.payments.arch.PurchaseRequest
+import com.appcoins.payments.arch.WalletData
 import kotlinx.coroutines.flow.flowOf
 
 class PaypalPaymentMethod internal constructor(
