+++ payments/payment-methods/paypal/src/main/java/com/appcoins/payment_method/paypal/PaypalPaymentMethodFactory.kt
@@ -1,11 +1,11 @@
 package com.appcoins.payment_method.paypal
 
-import com.appcoins.payment_manager.manager.PurchaseRequest
-import com.appcoins.payment_manager.payment.PaymentMethodFactory
-import com.appcoins.payment_manager.repository.broker.domain.PaymentMethodData
-import com.appcoins.payment_manager.wallet.WalletData
 import com.appcoins.payment_method.paypal.repository.PaypalRepository
-import com.appcoins.product_inventory.domain.ProductInfoData
+import com.appcoins.payments.arch.PaymentMethodData
+import com.appcoins.payments.arch.PaymentMethodFactory
+import com.appcoins.payments.arch.ProductInfoData
+import com.appcoins.payments.arch.PurchaseRequest
+import com.appcoins.payments.arch.WalletData
 import javax.inject.Inject
 import javax.inject.Singleton
