+++ payments/payment-methods/paypal/src/main/java/com/appcoins/payment_method/paypal/model/PaypalTransaction.kt
@@ -1,7 +1,7 @@
 package com.appcoins.payment_method.paypal.model
 
-import com.appcoins.payment_manager.transaction.Transaction
-import com.appcoins.payment_manager.transaction.TransactionStatus
+import com.appcoins.payments.arch.Transaction
+import com.appcoins.payments.arch.TransactionStatus
 import kotlinx.coroutines.flow.Flow
 
 class PaypalTransaction internal constructor(
